refactor: RestDocs 용어 및 양식 통일, 누락항목 추가, API 수정
- Umbrella 정보를 id로만 조회하도록 API 변경

+++ src/main/java/upbrella/be/rent/dto/request/RentUmbrellaByUserRequest.java
@@ -12,7 +12,7 @@ import lombok.NoArgsConstructor;
 public class RentUmbrellaByUserRequest {
 
     private String region;
-    private int storeId;
-    private int umbrellaId;
+    private long storeId;
+    private long umbrellaId;
     private String statusDeclaration;
 }
